Title: 평균점수 구하기, Time: 161ms, Memory: 12MB, Status: Passed - Codetree

+++ 250913/평균점수 구하기/print-average.java
@@ -3,10 +3,9 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         // Please write your code here.
-
     Scanner sc = new Scanner(System.in);
 
-    float sum=0;
+    double sum=0;
 
     for(int i=0;i<8;i++){
         double a= sc.nextDouble();
